refactor way how to work with tracking method
